Add CodeQL to Action workflows.
This change adds the default CodeQL analysis to our PR workflow.

It also turns on our default workflow to run on pushes. As a side-effect, this will ensure the build is ran if someone pushes directly to the repo with a PR, but the primary purpose is to configure CodeQL to run against the branch itself, and not just a PR. This is necessary for CodeQL to detect risk changes.

This PR will fail the CodeQL check due to some regex problems in the compiled docs output.

I have tested in my forked repo, and if we merge this anyway, then these alerts stick around in the `Security` tab in github, but subsequent PRs will not fail as CodeQL will recognize that we previously accepted these risks.

We need to figure out what to do with the doc site and fix it up, but I wanted to get this in so we get immediate feedback as we continue to write core code in this repo