Remove test that was breaking build.
This change uses `promise.reject` which, if uncaught, will fail the node eventloop and fail the tests.

My best guess is that timing on my mac are consistent enough, that this was failing and causing karma to bail out on the rest of the tests. Although I don't know why it works on github/windows... I also don't know why it used to work, very strange. Perhaps because of how we were polyfilling promises with babel before.

Looking at this test, we are testing standard promise stuff, and our Deferred class extends promise and doesn't override any of these functions. So deleting the test seems adequate.