Add engine-specific 'start' commands for running integration tests/examples (#95)
* Add engine-specific 'start' commands for running integration tests/examples.

In the root, you may now run `start-core`, `start-three`, or `start-babylon`.

These all run webpack dev server, but will open relevant tabs for the package you have selected. This is either integration tests, or both integration tests and examples.

More work is still required to integrate the integration test code/pages into webpack so they support hot reload

This change also enables source maps for non-production builds.

I've moved demo-credentials.js to the root of the repo. It is onerous for devs to insert what is likely the same cognito pool many places. I've created a fake API in webpack at /devConfig.json where webpack serves up the cognito pool and it is fetched via the samples. Ultimately I do thing we should serve up everything via webpack and probably HtmlWebpackPlugin

I tried to change this to a module with type: module in package.json. If you are reading this, do not attempt this until Karma 7! It doesn't seem like it can work without some really hacky stuff and I could not get it to work even then: Support for ESM config files (with "type": "module") karma-runner/karma#3677. My workaround is to change demo-credentials.js to be node-style instead of a ESM.

A release build won't build any of the examples. These don't work without webpack server because of file location and our devConfig endpoint, so I think it is confusing/misleading to add these to the build output. They wouldn't work anyway without cognito so I don't think this is so bad, but we need to figure out how to test these in CI and how to pull a sample cognito Id from github secrets/etc