Remove broken tests (#99)
See commit: 49e134c

This is the same thing again. We had this test copied 3 times(!) and the previous change only removed one of them.

To restate from before: These tests leave a dangling promise rejection, which end up failing the tests in a race condition. My best guess is that if the tests are ran on a fast enough machine that it passes as the rejection isn't observed on the event loop before all the tests have passed. This is only a guess though. This test doesn't seem to be testing anything useful anyway so removing them.

Leaving these tests in caused the tests to stop running on certain machines, including mine and Kris's.