Tune package exports and README improvements
I've updated the package.json for the published packages to provide both require(commonjs) and import(ESM) exports.

The idea here being that consumers of this package through NPM will default to `import` if they support it.
We generally expect applications using the library to bundle the source into their own applications, where they have full control over transpiling, chunking, polyfilling, etc according to their needs.

The .dist.js files are common.js files that are consumable directly in the browser. This can be used by applications that don't wish to use a bundler for their webapplication. It can also be used by any other consumer that is not compatible with ESM modules and must use commonjs for whatever reason.

I've also updated the README a bit to provide better guidance around usage, that is starting with it should be npm installed and not forked unless someone is actually contributing to the repo. I've mentioned our new sample application, but of course this needs to be expanded on when that is ready.