Add User Agent strings for Hosts package and Engine version.
The Sumerian Hosts now automatically add a version string for both the hosts package as well as the engine integration.

The Hosts version number is the git commit hash. This is set by github actions during a build, but using a local build will default to 'development'. We may want to switch this to release number which should be a simple change

The Engine version number identifies the version of Babylon or Three being used.

There are some unit tests
I validated by looking at the network inspector in the integration tests/demos for both lex(babylon-only) and polly

Example header from the babylon lex sample
`X-Amz-User-Agent: aws-sdk-js/2.1117.0 SumerianHosts-development babylonjs@4.2.2 promise`

The header does not get added to the presigned URLs we use for the MP3 download from Polly.
I don't think this is a problem, however, as we always send a second request to get the visemes which does have this header.